projet_streanlit/requirements.txt
os
leafmap.foliumap 
streamlit 
reportlab.pdfgen 
io 
PIL
streamlit_option_menu
folium
streamlit_folium
urllib.error
pandas
geopandas
shapely.wkb
branca.colormap
folium.plugins
selenium 
selenium.webdriver.firefox.options
time
tempfile
leafmap 
geopy.distance
rasterio
cv2
imageio
numpy 
plotly.express
plotly.io
matplotlib.pyplot
base64